Renamed the JDL app file
