+++ lib/core/abstract_jdl_application.js
