+++ lib/core/jdl_monolith_application.js
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-const AbstractJDLApplication = require('./jdl_application');
+const AbstractJDLApplication = require('./abstract_jdl_application');
 const ApplicationTypes = require('./jhipster/application_types');
 const ApplicationOptions = require('./jhipster/application_options');
