+++ lib/core/jdl_object.js
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-const AbstractJDLApplication = require('./jdl_application');
+const AbstractJDLApplication = require('./abstract_jdl_application');
 const JDLEntity = require('./jdl_entity');
 const JDLEnums = require('./jdl_enums');
 const JDLEnum = require('./jdl_enum');
