+++ lib/export/jhipster_application_exporter.js
@@ -19,7 +19,7 @@
 
 const fs = require('fs');
 const path = require('path');
-const AbstractJDLApplication = require('../core/jdl_application');
+const AbstractJDLApplication = require('../core/abstract_jdl_application');
 const FileUtils = require('../utils/file_utils');
 
 module.exports = {
